Revert "Rm postship transaction"
